do meltdown

+++ assignment4/meltdown-template.c
@@ -92,7 +92,7 @@ static int cc_receive() {
     u64 time;
     void *cur_adrs;
     int ret=-1;
-    for (int i=0; i<cc_buf_size/4096;i++){
+    for (int i=0; i< (int)cc_buf_size/4096;i++){
         asm volatile("mfence\n");
         cur_adrs=&cc_buffer[i*512+i];
         time = my_rdtsc();
